requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
cartesia==1.0.14
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.2
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
iterators==0.2.0
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
propcache==0.2.0
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yarl==1.15.5
